+++ KDIC/PlayerViewController.swift
@@ -22,7 +22,7 @@ class PlayerViewController: UIViewController {
     }
     
     func togglePlayer() {
-        if KdicPlayer.sharedInstance.currentlyPlaying() {
+        if KDICPlayer.sharedInstance.isPlaying {
             pauseRadio()
             playerButton.setImage(UIImage(named: "playButton"), forState: UIControlState.Normal)
             
@@ -33,22 +33,12 @@ class PlayerViewController: UIViewController {
     }
     
     func playRadio() {
-        KdicPlayer.sharedInstance.play()
+        KDICPlayer.sharedInstance.play()
     }
     
     func pauseRadio() {
-        KdicPlayer.sharedInstance.pause()
+        KDICPlayer.sharedInstance.pause()
         
     }
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
-        // Get the new view controller using segue.destinationViewController.
-        // Pass the selected object to the new view controller.
-    }
-    */
 
 }
