+++ KDIC/ScheduleTableCell.swift
@@ -6,8 +6,4 @@ class ScheduleTableCell: UITableViewCell {
     @IBOutlet weak var showName: UILabel!
     @IBOutlet weak var showTime: UILabel!
     
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        
-    }
 }
